decode bitwise operations and shifts
this comes with a large refactor of the main instruction decoding
function. previously, it had heuristics inside to determine which order
the various decoded parts (mod_rm, reg, immediate) should go in the
final instruction. there proved to be too much variance for this to
work. now there are two parts to the decode:

1. decode the parts (mod_rm, reg, immediate)
2. decide how to order them

this is a lot more flexible and lets me handle the weird instructions
more flexibly.

decode.zig could probably still use another refactor pass.