introduce Prefixes to Instructions and SrcDsts
8086 instructions and operands can have various prefixes attached to
them, e.g.

   rep movsb
   lock not byte [...]
   mov al, cs:[]

this commit adds a new Prefix type that gets put on to the Instruction
and SrcDst structs. The old size field was merged into the prefixes
field. This is nice since most instructions don't actually *need* size
prefixes to begin with, so we can leave them out now instead of the
weird size = .unspecified workaround we had.

it might be worth considering a more data-oriented approach for this,
since the prefix fields are probably mostly empty space. but for now
they are only two elements of a 2 byte enum (so 4 bytes in total), so
it's probably not hurting too much.